TasksPage PR fix
