+++ src/components/TasksPage/TasksPage.jsx
@@ -1,11 +1,6 @@
 import React from "react";
-import { Routes, Route, NavLink } from "react-router-dom";
+import { NavLink, Outlet } from "react-router-dom";
 import styles from "./TasksPage.module.css";
-
-const AllTasks = () => <h2>Все задачи</h2>;
-const TodayTasks = () => <h2>Задачи на сегодня</h2>;
-const CompletedTasks = () => <h2>Выполненные</h2>;
-const DeleteTasks = () => <h2>Удаленные</h2>;
 
 export const TasksPage = () => {
   return (
@@ -14,7 +9,7 @@ export const TasksPage = () => {
         <ul>
           <li>
             <NavLink
-              to="/tasks/all"
+              to="all"
               className={({ isActive }) => (isActive ? styles.active : "")}
             >
               Все задачи
@@ -22,7 +17,7 @@ export const TasksPage = () => {
           </li>
           <li>
             <NavLink
-              to="/tasks/today"
+              to="today"
               className={({ isActive }) => (isActive ? styles.active : "")}
             >
               Задачи на сегодня
@@ -30,7 +25,7 @@ export const TasksPage = () => {
           </li>
           <li>
             <NavLink
-              to="/tasks/completed"
+              to="completed"
               className={({ isActive }) => (isActive ? styles.active : "")}
             >
               Выполненные
@@ -38,7 +33,7 @@ export const TasksPage = () => {
           </li>
           <li>
             <NavLink
-              to="/tasks/deleted"
+              to="deleted"
               className={({ isActive }) => (isActive ? styles.active : "")}
             >
               Удаленные
@@ -47,12 +42,7 @@ export const TasksPage = () => {
         </ul>
       </div>
       <div className={styles.content}>
-        <Routes>
-          <Route path="all" element={<AllTasks />} />
-          <Route path="today" element={<TodayTasks />} />
-          <Route path="completed" element={<CompletedTasks />} />
-          <Route path="deleted" element={<DeleteTasks />} />
-        </Routes>
+        <Outlet />
         <h1>Будущие задачи</h1>
       </div>
     </main>
