+++ src/index.css
@@ -56,17 +56,21 @@ body {
   --text-color: #000;
   --window-background-color: #fff;
   --side-panel-background-color: #fff;
+  --hover-background-color: #e0e0e0;
+  --active-background-color: #ccc;
   margin: 0;
   padding: 0;
 }
 
 body.dark {
   font-family: "Inter";
-  --header-background-color: rgba(0, 0, 0, 0.65);
+  --header-background-color: rgba(0, 0, 0, 0.8);
   --header-text-color: #fff;
-  --main-background-color: rgba(0, 0, 0, 0.65);
+  --main-background-color: rgba(0, 0, 0, 0.8);
   --main-text-color: #fff;
   --text-color: #fff;
-  --window-background-color: rgba(0, 0, 0, 0.65);
-  --side-panel-background-color: rgba(0, 0, 0, 0.65);
+  --window-background-color: rgba(0, 0, 0, 0.8);
+  --side-panel-background-color: rgba(0, 0, 0, 0);
+  --hover-background-color: rgba(0, 0, 0, 0.8);
+  --active-background-color: rgba(0, 0, 0, 0.8);
 }
